requirements.txt
fastapi
uvicorn
google-generativeai
pygithub
python-dotenv
aiohttp
requests